Use type float32 instead of float64 for regularization weights.

+++ cpp/subprojects/boosting/include/mlrl/boosting/rule_evaluation/regularization_no.hpp
@@ -13,7 +13,7 @@ namespace boosting {
     class NoRegularizationConfig final : public IRegularizationConfig {
         public:
 
-            float64 getWeight() const override;
+            float32 getWeight() const override;
     };
 
 }
