+++ cpp/subprojects/boosting/include/mlrl/boosting/rule_evaluation/rule_evaluation_decomposable_partial_dynamic.hpp
@@ -20,9 +20,9 @@ namespace boosting {
 
             const float32 exponent_;
 
-            const float64 l1RegularizationWeight_;
+            const float32 l1RegularizationWeight_;
 
-            const float64 l2RegularizationWeight_;
+            const float32 l2RegularizationWeight_;
 
         public:
 
@@ -44,8 +44,8 @@ namespace boosting {
              *                                  scores to be predicted by rules
              */
             DecomposableDynamicPartialRuleEvaluationFactory(float32 threshold, float32 exponent,
-                                                            float64 l1RegularizationWeight,
-                                                            float64 l2RegularizationWeight);
+                                                            float32 l1RegularizationWeight,
+                                                            float32 l2RegularizationWeight);
 
             std::unique_ptr<IRuleEvaluation<DenseDecomposableStatisticVector<float64>>> create(
               const DenseDecomposableStatisticVector<float64>& statisticVector,
