+++ cpp/subprojects/boosting/src/mlrl/boosting/binning/label_binning_no.cpp
@@ -15,8 +15,8 @@ namespace boosting {
 
     std::unique_ptr<IDecomposableRuleEvaluationFactory>
       NoLabelBinningConfig::createDecomposableCompleteRuleEvaluationFactory() const {
-        float64 l1RegularizationWeight = l1RegularizationConfig_.get().getWeight();
-        float64 l2RegularizationWeight = l2RegularizationConfig_.get().getWeight();
+        float32 l1RegularizationWeight = l1RegularizationConfig_.get().getWeight();
+        float32 l2RegularizationWeight = l2RegularizationConfig_.get().getWeight();
         return std::make_unique<DecomposableCompleteRuleEvaluationFactory>(l1RegularizationWeight,
                                                                            l2RegularizationWeight);
     }
@@ -24,8 +24,8 @@ namespace boosting {
     std::unique_ptr<ISparseDecomposableRuleEvaluationFactory>
       NoLabelBinningConfig::createDecomposableFixedPartialRuleEvaluationFactory(float32 outputRatio, uint32 minOutputs,
                                                                                 uint32 maxOutputs) const {
-        float64 l1RegularizationWeight = l1RegularizationConfig_.get().getWeight();
-        float64 l2RegularizationWeight = l2RegularizationConfig_.get().getWeight();
+        float32 l1RegularizationWeight = l1RegularizationConfig_.get().getWeight();
+        float32 l2RegularizationWeight = l2RegularizationConfig_.get().getWeight();
         return std::make_unique<DecomposableFixedPartialRuleEvaluationFactory>(
           outputRatio, minOutputs, maxOutputs, l1RegularizationWeight, l2RegularizationWeight);
     }
@@ -33,8 +33,8 @@ namespace boosting {
     std::unique_ptr<ISparseDecomposableRuleEvaluationFactory>
       NoLabelBinningConfig::createDecomposableDynamicPartialRuleEvaluationFactory(float32 threshold,
                                                                                   float32 exponent) const {
-        float64 l1RegularizationWeight = l1RegularizationConfig_.get().getWeight();
-        float64 l2RegularizationWeight = l2RegularizationConfig_.get().getWeight();
+        float32 l1RegularizationWeight = l1RegularizationConfig_.get().getWeight();
+        float32 l2RegularizationWeight = l2RegularizationConfig_.get().getWeight();
         return std::make_unique<DecomposableDynamicPartialRuleEvaluationFactory>(
           threshold, exponent, l1RegularizationWeight, l2RegularizationWeight);
     }
@@ -42,8 +42,8 @@ namespace boosting {
     std::unique_ptr<INonDecomposableRuleEvaluationFactory>
       NoLabelBinningConfig::createNonDecomposableCompleteRuleEvaluationFactory(const Blas& blas,
                                                                                const Lapack& lapack) const {
-        float64 l1RegularizationWeight = l1RegularizationConfig_.get().getWeight();
-        float64 l2RegularizationWeight = l2RegularizationConfig_.get().getWeight();
+        float32 l1RegularizationWeight = l1RegularizationConfig_.get().getWeight();
+        float32 l2RegularizationWeight = l2RegularizationConfig_.get().getWeight();
         return std::make_unique<NonDecomposableCompleteRuleEvaluationFactory>(l1RegularizationWeight,
                                                                               l2RegularizationWeight, blas, lapack);
     }
@@ -53,8 +53,8 @@ namespace boosting {
                                                                                    uint32 minOutputs, uint32 maxOutputs,
                                                                                    const Blas& blas,
                                                                                    const Lapack& lapack) const {
-        float64 l1RegularizationWeight = l1RegularizationConfig_.get().getWeight();
-        float64 l2RegularizationWeight = l2RegularizationConfig_.get().getWeight();
+        float32 l1RegularizationWeight = l1RegularizationConfig_.get().getWeight();
+        float32 l2RegularizationWeight = l2RegularizationConfig_.get().getWeight();
         return std::make_unique<NonDecomposableFixedPartialRuleEvaluationFactory>(
           outputRatio, minOutputs, maxOutputs, l1RegularizationWeight, l2RegularizationWeight, blas, lapack);
     }
@@ -63,8 +63,8 @@ namespace boosting {
       NoLabelBinningConfig::createNonDecomposableDynamicPartialRuleEvaluationFactory(float32 threshold,
                                                                                      float32 exponent, const Blas& blas,
                                                                                      const Lapack& lapack) const {
-        float64 l1RegularizationWeight = l1RegularizationConfig_.get().getWeight();
-        float64 l2RegularizationWeight = l2RegularizationConfig_.get().getWeight();
+        float32 l1RegularizationWeight = l1RegularizationConfig_.get().getWeight();
+        float32 l2RegularizationWeight = l2RegularizationConfig_.get().getWeight();
         return std::make_unique<NonDecomposableDynamicPartialRuleEvaluationFactory>(
           threshold, exponent, l1RegularizationWeight, l2RegularizationWeight, blas, lapack);
     }
